% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_lnm_helper_functions.R
\name{copy_lnm_attributes}
\alias{copy_lnm_attributes}
\title{Given a torch_lnm module, mod, with params_vec, this copies all the attributes of mod 
into a new torch_lnm_stepwise module (except for params_free and params_free_sizes) which 
is instantiated during pruning/stepping up.}
\usage{
copy_lnm_attributes(mod, params_vec)
}
\arguments{
\item{params_vec}{params_vec which belongs to the torch_lnm module. If name of module is mod,
"mod$params_vec" should be input here.}

\item{torch_lnm}{torch_lnm module which is undergoing pruning/stepping up}
}
\value{
torch_lnm_stepwise module with the exact same attributes as the original torch_lnm module
}
\description{
Given a torch_lnm module, mod, with params_vec, this copies all the attributes of mod 
into a new torch_lnm_stepwise module (except for params_free and params_free_sizes) which 
is instantiated during pruning/stepping up.
}
