% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_rnm_helper_functions.R
\name{copy_rnm_attributes}
\alias{copy_rnm_attributes}
\title{Given a torch_rnm module, mod, with params_vec, this copies all the attributes of mod 
into a new torch_rnm_stepwise module (except for params_free and params_free_sizes) which 
is instantiated during stepping up/stepping down/pruning.}
\usage{
copy_rnm_attributes(mod, params_vec)
}
\arguments{
\item{params_vec}{params_vec which belongs to the torch_enm module. If name of module is mod,
"mod$params_vec" should be input here.}

\item{torch_rnm}{torch_rnm module which is undergoing stepping down/stepping up/pruning}
}
\value{
torch_rnm_stepwise module with the exact same attributes as the original torch_enm module
}
\description{
Given a torch_rnm module, mod, with params_vec, this copies all the attributes of mod 
into a new torch_rnm_stepwise module (except for params_free and params_free_sizes) which 
is instantiated during stepping up/stepping down/pruning.
}
