% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_lnm_helper_functions.R
\name{df_to_tensor}
\alias{df_to_tensor}
\title{Prepares a dataframe for the lnm/rnm model. This function
first converts the variables to a design matrix, then centers
it, and lastly converts it to a torch tensor}
\usage{
df_to_tensor(df, dtype = NULL, device = NULL)
}
\arguments{
\item{df}{data frame}

\item{dtype}{data type of the resulting tensor}

\item{device}{device to store the resulting tensor on}
}
\value{
Torch tensor of scaled and processed data
}
\description{
#' Example: df_to_tensor(mtcars)
}
\details{
Output: torch tensor with columns corresponding to mtcars variables

Explanation: Each variable mtcars is centered and the resulting 
dataframe is first converted to a torch tensor before it is
returned.
}
\seealso{
[torch::torch_tensor()], [stats::model.matrix()]
}
