% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{diagonal_idx}
\alias{diagonal_idx}
\title{For an adjacency matrix A, indicates the indices of 
the entries of the diagonal of A inside vec(A)}
\usage{
diagonal_idx(n)
}
\arguments{
\item{n}{dimension of the matrix}
}
\value{
array of indices
}
\description{
Example: diagonal_idx(3) 
Output: 1 4 6
Explanation: the entries referred to by the indices
correspond to (1,1), (2,2) and (3,3)
entries of the symmmetric matrix A 
which are the 1, 4 and 6th entries
of vec(A) when read from top to bottom
}
