% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_lnm_helper_functions.R
\name{find_correct_indices}
\alias{find_correct_indices}
\title{Returns the indices of an reference array which are found in an input array. If none are found,
Null is returned}
\usage{
find_correct_indices(reference_array, input_array)
}
\arguments{
\item{x}{reference array}

\item{n}{input array}
}
\value{
vector of indices denoting positions of entries in reference array found
in input array
}
\description{
Example: find_correct_indices(c(0,2,3),c(4,5,6,0))
}
\details{
Output: 1
 

Explanation: The value 0 in the reference array (first array) is found in the second array and the 
position of 0 in the reference array is 1. So 1 is returned.
}
