% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_lnm_helper_functions.R
\name{flip_one_1_to_0}
\alias{flip_one_1_to_0}
\title{Given a tensor of 1s and 0s, it returns a list of the possible tensors after one of the 1s 
is now set to 0}
\usage{
flip_one_1_to_0(tensor)
}
\arguments{
\item{tensor}{torch tensor of 1s and 0s}
}
\value{
list of possible torch tensors
}
\description{
Example: flip_one_1_to_0(torch_tensor(c(1,0,0)))
Output: [[1]]
       torch_tensor
       0
       0
       0
       [ CPUFloatType{3} ]
       
Explanation: There is only one possibility from torch_tensor(c(1,0,0)), namely
torch_tensor(c(0,0,0))
}
