% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{format_lambda_mat}
\alias{format_lambda_mat}
\title{Given the lambda matrix belonging to an lnm/rnm, and the type of matrix needed specified,
it returns the matrix either representing the free parameters of the lambda matrix
of the matrix of the starting values of the lambda matrix.}
\usage{
format_lambda_mat(mat, type = NULL)
}
\arguments{
\item{matrix}{lambda matrix of}

\item{string}{"free" or "value"}
}
\value{
binary matrix which represents the free parameters of the lambda matrix
or the matrix which contains the starting values of the lambda matrix.
}
\description{
Example: lambda <- matrix(0,10,2)
         lambda[c(1,2,4,6,7),1] <- 1 
         lambda[c(3,5,8,9,10),2] <- 1 
         format_lambda_mat(lambda,type = "free")
         format_lambda_mat(lambda,type = "value")
         
Output: 
       [,1] [,2]
 [1,]    0    0
 [2,]    1    0
 [3,]    0    0
 [4,]    1    0
 [5,]    0    1
 [6,]    1    0
 [7,]    1    0
 [8,]    0    1
 [9,]    0    1
 [10,]   0    1
 
 
       [,1] [,2]
 [1,]     1    0
 [2,]     0    0
 [3,]     0    1
 [4,]     0    0
 [5,]     0    0
 [6,]     0    0
 [7,]     0    0
 [8,]     0    0
 [9,]     0    0
 [10,]    0    0
 
       
Explanation: There is only one possibility from torch_tensor(c(0,1,1)), namely
torch_tensor(c(1,1,1))
}
