% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{format_omega_theta}
\alias{format_omega_theta}
\title{Format omega_theta matrix for rnm module}
\usage{
format_omega_theta(mat)
}
\arguments{
\item{mat}{omega_theta matrix}
}
\value{
a list containing the following: 1)rows of nonzero partial correlations,
2)cols of nonzero partial correlations and 3)initial values
}
\description{
Given the omega_theta matrix of an existing rnm model either as an R matrix or as
a torch tensor, the function returns the omega_theta matrix in a format for the training
of a new rnm model on another set of data but with the same omega_theta matrix
}
