% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{get_alternative_update}
\alias{get_alternative_update}
\title{Scales the current parameter estimates by a specified scale factor in the case
the current estimates lead to a covariance matrix which is not positive definite. 
The default value of the scaling factor is 0.9}
\usage{
get_alternative_update(current_params, scale_factor = 0.9)
}
\arguments{
\item{tensor}{current parameters of the lnm/rnm model}

\item{numeric}{factor to scale the current parameters}
}
\value{
torch tensor containing the new parameters
}
\description{
Example: get_alternative_update(torch_tensor(c(0,1,1)))
Output: 
       torch_tensor
       0
       0.9
       0.9
       [ CPUFloatType{3} ]
       
Explanation: The individual values are scaled by a factor of 0.9
}
