% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_magna_helper_functions.R
\name{get_corMats_as_tensors}
\alias{get_corMats_as_tensors}
\title{For a list of correlation matrices (as R matrices) provided, returns the same list
but with matrices as torch tensors}
\usage{
get_corMats_as_tensors(corMats)
}
\arguments{
\item{corMats}{list of matrices as base R matrices}
}
\value{
list of matrices as torch tensors
}
\description{
Example:
}
\details{
corMats <- list(
matrix(
 c(1, 0.31, -0.09,
   0.31, 1, -0.23,
   -0.09, -0.23, 1),
 nrow = 3, ncol = 3),
matrix(c(
 1, 0.27, -0.01,
 0.27, 1, -0.18,
 -0.01, -0.18, 1),
 nrow = 3, ncol = 3),
matrix(c(
 1, 0.23, -0.01,
 0.23, 1, -0.25,
 -0.01, -0.25, 1),
 nrow = 3, ncol = 3)
)
get_corMats_as_tensors(corMats) 
Output: Same list but matrices are now torch tensors
}
