% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_magna_helper_functions.R
\name{get_mean_corr}
\alias{get_mean_corr}
\title{For a list of correlation matrices (as torch tensors), the dimension of the network (n) and the number of studies (m) provided, 
returns the mean vector of the correlations present in the network}
\usage{
get_mean_corr(corMats, n, m)
}
\arguments{
\item{corMats}{list of torch tensors}

\item{n}{dimension of the network}

\item{m}{number of studies}
}
\value{
mean vector of correlations present in the network
}
\description{
Example:
}
\details{
corMats <- get_corMats_as_tensors(corMats) 
get_mean_corr(corMats,3,3) 

Output: Mean vector of the correlations present in the network

Explanation: Extracts and sums the correlations from the matrices by using the dimension network and then divides
the derived sum over the number of studies
}
