% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_magna_helper_functions.R
\name{get_network_dim}
\alias{get_network_dim}
\title{For a list of correlation matrices (as torch tensors) provided, returns the dimension of the network
which is assumed to be the same for each correlation matrix.}
\usage{
get_network_dim(corMats)
}
\arguments{
\item{corMats}{list of torch tensors}
}
\value{
dimension of network
}
\description{
Example:
}
\details{
corMats <- list(
matrix(
 c(1, 0.31, -0.09,
   0.31, 1, -0.23,
   -0.09, -0.23, 1),
 nrow = 3, ncol = 3),
matrix(c(
 1, 0.27, -0.01,
 0.27, 1, -0.18,
 -0.01, -0.18, 1),
 nrow = 3, ncol = 3),
matrix(c(
 1, 0.23, -0.01,
 0.23, 1, -0.25,
 -0.01, -0.25, 1),
 nrow = 3, ncol = 3)
)
corMats <- get_corMats_as_tensors(corMats) 

get_network_dim(corMats) 
Output: 3

Explanation: the dimension of the network is 3
}
