% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_lnm_helper_functions.R
\name{get_p_value}
\alias{get_p_value}
\title{Returns the p-value of a 2-tailed Z test}
\usage{
get_p_value(x)
}
\arguments{
\item{x}{Z-statistic observed}
}
\value{
p-value from a 2-tailed Z test
}
\description{
Example: get_p_value(1.96)
Output: 0.04999579
}
\details{
Explanation: Z-statistic of 1.96 roughly corresponds to a p-value of 0.05 from a 2-tailed Z test.
}
