% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_magna_helper_functions.R
\name{get_sample_corr_of_corr}
\alias{get_sample_corr_of_corr}
\title{For a list of correlation matrices (as torch tensors), the vector of correlation means,
the number of omega params, the dimension of the network (n) and the number of studies (m) provided, 
returns the sample correlation matrix of the correlations}
\usage{
get_sample_corr_of_corr(corMats, mean, omega_params_length, n, m)
}
\arguments{
\item{corMats}{list of torch tensors}

\item{mean}{mean vector of correlations, by default NULL}

\item{omega_params_length}{number of omega parameters under consideration}

\item{n}{dimension of the network}

\item{m}{number of studies}
}
\value{
sample correlation matrix of correlations
}
\description{
Example:
}
\details{
get_sample_corr_of_corr(corMats,mean = NULL, omega_params_length = 3,3,3) 
Output: sample correlation matrix of the correlations

Explanation: Uses information provided to calculate sample correlations
}
