% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_rnm_helper_functions.R
\name{get_saturated_omega_theta_free}
\alias{get_saturated_omega_theta_free}
\title{Given a number n (number of observations in dataset), returns a the torch tensor
to represent the free params of the omega delta matrix assuming a saturated model}
\usage{
get_saturated_omega_theta_free(n)
}
\arguments{
\item{n}{}
}
\value{
tensor
}
\description{
Example: get_saturated_omega_theta_free(3)
Output: 
       torch_tensor
       0
       1
       1
       0
       1
       0
       [ CPUFloatType{6} ]
       
Explanation: There are in total 3 free params inside the matrix, the first 2
belong to the first column in the lower triangle of the matrix and the last
one belongs to the second column
}
