% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_magna_helper_functions.R
\name{kron}
\alias{kron}
\title{Outputs the Kronecker product of A and B, 2 torch tensors}
\usage{
kron(A, B)
}
\arguments{
\item{A}{left torch tensor}

\item{B}{right torch tensor}
}
\value{
Kronecker product of A and B
}
\description{
Example:
kron(torch_eye(2),torch_eye(2))
}
\details{
Output: 
1  0  0  0
0  1  0  0
0  0  1  0
0  0  0  1

Explanation: Recall how the Kronecker product works, where each entry in the left matrix is multiplied 
with the right matrix - these individual "sub" products each form a block matrix in the new output matrix
}
