% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_lnm_helper_functions.R
\name{lnm_prune}
\alias{lnm_prune}
\title{Pruning fortorch_lnm/torch_lnm_stepwise module}
\usage{
lnm_prune(mod)
}
\arguments{
\item{mod}{original torch_lnm/torch_lnm_stepwise module}
}
\value{
new refitted model with only significant partial correlations
}
\description{
Given an  model, it iteratively removes insignificant partial correlations and refits models until
all partial correlations are found to be significant. Thereafter ,it returns this model
with only significant partial correlations.
}
