% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_lnm_helper_functions.R
\name{lnm_prune_helper}
\alias{lnm_prune_helper}
\title{Helper function for lnm_prune. Given an lnm model, it removes all partial correlations
found to be insignificant from the model and refits the model without these partial correlations.
Thereafter, it returns the updated model.}
\usage{
lnm_prune_helper(mod, batch_size = NULL)
}
\arguments{
\item{mod}{original torch_lnm/torch_lnm_stepwise module}
}
\value{
new refitted model with insignificant partial correlations removed
}
\description{
Helper function for lnm_prune. Given an lnm model, it removes all partial correlations
found to be insignificant from the model and refits the model without these partial correlations.
Thereafter, it returns the updated model.
}
