% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_lnm_helper_functions.R
\name{lnm_stepup}
\alias{lnm_stepup}
\title{Step Up function for torch_lnm/torch_lnm_stepwise module}
\usage{
lnm_stepup(mod, criterion = "BIC", gamma = 0.5)
}
\arguments{
\item{mod}{original torch_lnm/torch_lnm_stepwise module undergoing stepping up}

\item{criterion}{string indicating criterion to use - "AIC", "BIC or "EBIC"}

\item{gamma}{(Optional) value of gamma if "EBIC" is used. Default value is 0.5.}
}
\value{
either the original torch_lnm/torch_lnm_stepwise module or 
a torch_lnm_stepwise module with a best criterion score after stepping down
}
\description{
Performs stepping up to give the best model based on the selected criterion score
At the end of each iteration, one of the originally zero partial correlations will be
turned to non-zero and this corresponds to the model with the best model fit based on criterion score
out of all possibilities (number of zero partial correlations at the start). 
If the models fit at the end of 1 iteration with 1 partial correlation set to non-zero each 
do not lead to a better criterion score, the model found at the end of the previous iteration is returned.
}
\details{
The process terminates if the model is saturated (all partial correlations are non-zero) at the end of the iteration.
}
