% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_lnm_helper_functions.R
\name{lower_triangle_idx}
\alias{lower_triangle_idx}
\title{For a symmetric matrix A, indicates the indices of 
the entries of the lower triangle of A inside vec(A)}
\usage{
lower_triangle_idx(n)
}
\arguments{
\item{n}{dimension of the adjacency matrix}
}
\value{
array of indices
}
\description{
Example: lower_triangle_idx(3) 
Output: 1 2 3 5 6 9
Explanation: the entries referred to by the indices
correspond to (1,1), (2,1), (3,1), (2,2), (3,2), (3,3) 
entries of the symmetric matrix A 
which are the 1, 2, 3, 5, 6 and 9th entries
of vec(A) when read from top to bottom
}
