% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_magna_helper_functions.R
\name{random_effects_MAGNA_prune}
\alias{random_effects_MAGNA_prune}
\title{Prunes a given random effects MAGNA model and refits the pruned model, assuming standard errors for all
initial omega parameters are available}
\usage{
random_effects_MAGNA_prune(
  mod,
  lrate = 0.05,
  maxit = 5000,
  verbose = TRUE,
  tol = 1e-20
)
}
\arguments{
\item{mod}{Random effects MAGNA model}

\item{lrate}{learning rate for refitting the model, default 0.05}

\item{maxit}{maximum number of iterations for refitting the model, default 5000}

\item{verbose}{indicate whether loss function value should be printed for each iteration of optimization, default TRUE}

\item{tol}{threshold needed to be met for convergence, default 1e-20}
}
\value{
Random effects MAGNA model mod with non-significant omega parameters removed
}
\description{
Example:
random_effects_MAGNA_prune(mod)
}
\details{
Output: Random effects MAGNA model mod with non-significant omega parameters removed
}
