% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_rnm_helper_functions.R
\name{rnm_prune}
\alias{rnm_prune}
\title{Pruning function for torch_rnm/torch_rnm_stepwise module}
\usage{
rnm_prune(mod, batch_size = NULL)
}
\arguments{
\item{mod}{original torch_rnm/torch_rnm_stepwise module}
}
\value{
new refitted model with only significant partial correlations
}
\description{
Given an r nm model, it iteratively removes insignificant partial correlations and refits models until
all partial correlations are found to be significant. Thereafter ,it returns this model
with only significant partial correlations.
}
