% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_rnm_helper_functions.R
\name{rnm_prune_helper}
\alias{rnm_prune_helper}
\title{Helper function for rnm_prune. Given an rnm model, it removes all partial correlations
found to be insignificant from the model and refits the model without these partial correlations.
Thereafter, it returns the updated model.}
\usage{
rnm_prune_helper(mod)
}
\arguments{
\item{mod}{original torch_rnm/torch_rnm_stepwise module}
}
\value{
new refitted model with insignificant partial correlations removed
}
\description{
Helper function for rnm_prune. Given an rnm model, it removes all partial correlations
found to be insignificant from the model and refits the model without these partial correlations.
Thereafter, it returns the updated model.
}
