% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_rnm_helper_functions.R
\name{rnm_stepdown}
\alias{rnm_stepdown}
\title{NOT RECOMMENDED FOR RNM DUE TO LONG EXECUTION TIMES}
\usage{
rnm_stepdown(mod, criterion = "BIC", gamma = 0.5)
}
\arguments{
\item{mod}{original torch_rnm/torch_rnm_stepwise module undergoing stepping up}

\item{criterion}{string indicating criterion to use - "AIC", "BIC or "EBIC"}

\item{gamma}{(Optional) value of gamma if "EBIC" is used. Default value is 0.5.}
}
\value{
either the original torch_rnm/torch_rnm_stepwise module or
}
\description{
Step Up function for torch_rnm/torch_rnm_stepwise module
}
\details{
Performs stepping up to give the best model based on the selected criterion score
At the end of each iteration, one of the originally zero partial correlations will be
turned to non-zero and this corresponds to the model with the best model fit based on criterion score
corresponds to the model with the best model fit based on criterion score 
out of all possibilities (number of zero partial correlations at the start). 
If the models fit at the end of 1 iteration with 1 partial correlation set to non-zero each 
do not lead to a better criterion score, the model found at the end of the previous iteration is returned.

The process terminates if the model is saturated (all partial correlations are non-zero) at the end of the iteration.
}
