% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_rnm_helper_functions.R
\name{rnm_stepdown_find_alt_models}
\alias{rnm_stepdown_find_alt_models}
\title{NOT RECOMMENDED FOR RNM DUE TO LONG EXECUTION TIMES}
\usage{
rnm_stepdown_find_alt_models(
  mod,
  criterion = "BIC",
  gamma = 0.5,
  batch_size = NULL
)
}
\arguments{
\item{mod}{original torch_rnm/torch_rnm_stepwise module undergoing stepping up}

\item{criterion}{string indicating criterion to use - "AIC", "BIC or "EBIC"}

\item{gamma}{(Optional) value of gamma if "EBIC" is used. Default value is 0.5.}
}
\value{
either the original torch_rnm/torch_rnm_stepwise module or 
a torch_rnm_stepwise module with a better criterion score
}
\description{
#' Helper function for rnm_stepup. Given an original torch_rnm/torch_rnm_stepwise module, the function fits all possible models
during one iteration of the stepping up process and returns either the best model based on criterion score
after this fitting process or the original torch_rnm/torch_rnm_stepwise module if all new models give a
worse model fit based on criterion score. A possible model here is a model with one of the
originally zero partial correlations now being set to non-zero.
}
