% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_magna_helper_functions.R
\name{strict_duplication_matrix}
\alias{strict_duplication_matrix}
\title{Outputs a strict duplication matrix as a torch tensor for any requested dimension n}
\usage{
strict_duplication_matrix(n)
}
\arguments{
\item{n}{dimension as requested}
}
\value{
strict duplication matrix of dimension n as a torch tensor
}
\description{
Example:
strict_duplication_matrix(2)
}
\details{
Output: 
 0
 1
 1
 0

Explanation: Left multiplying the strict duplication matrix for any vechs(X) gives vech(X)
}
