% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_lnm_helper_functions.R
\name{strict_lower_triangle_idx}
\alias{strict_lower_triangle_idx}
\title{For a symmetric matrix A, indicates the indices of 
the entries of the strict lower triangle of A inside vec(A)}
\usage{
strict_lower_triangle_idx(n)
}
\arguments{
\item{n}{dimension of the adjacency matrix}
}
\value{
array of indices
}
\description{
Example: strict_lower_triangle_idx(3) 
Output: 2 3 6
Explanation: the entries referred to by the indices
correspond to (2,1), (3,1) and (3,2) 
entries of the symmetric matrix A 
which are the 2, 3 and 6th entries
of vec(A) when read from top to bottom
}
