% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_lnm_helper_functions.R
\name{tensor_lnm_stepwise}
\alias{tensor_lnm_stepwise}
\title{Torch lnm module created during stepping down/stepping up/pruning}
\usage{
tensor_lnm_stepwise(params_vec)
}
\arguments{
\item{params_vec}{param values for the initializing of the module}
}
\value{
A `torch_sem` object, which is an `nn_module` (torch object)
}
\description{
Function for creating an lnm model
}
\details{
This function instantiates a torch object for computing the model-implied covariance matrix
based on an lnm model during the stepping down/stepping up/pruning processes. The methods are identical 
to torch_lnm and information about these methods can be looked up under torch_lnm
}
\section{Methods}{


## `$forward()`
Compute the model-implied covariance matrix.
Don't use this; `nn_modules` are callable, so access this method by calling
the object itself as a function, e.g., `my_torch_lnm()`.
In the forward pass, we apply constraints to the parameter vector, and we
create matrix views from it to compute the model-implied covariance matrix.

### Value
A `torch_tensor` of the model-implied covariance matrix



## `$loglik()`
Multivariate normal log-likelihood of the data.


### Value
Log-likelihood value (torch scalar)



## `$lasso_loss()`
Returns lasso loss which is -2 times Multivariate normal log-likelihood of the data + Penalty.

### Arguments
- `v` hyperparameter which controls for the penalty term inside the lasso loss function

### Value
lasso loss value (torch scalar)



## `$inverse_Hessian()`
Compute and return the asymptotic covariance matrix of the parameters with
respect to the loss function, which is limited to the -2 times the log-likelihood function


### Value
A `torch_tensor`, representing the ACOV of the free parameters



## `$fit()`
Fit a torch_rnm model using the default maximum likelihood objective.
This function uses the Adam optimizer to estimate the parameters of a torch_rnm

### Arguments
- `lrate` (Optional) learning rate of the Adam optimizer. Default is 0.05.
- `maxit` (Optional) maximum number of epochs to train the model. Default is 5000.
- `verbose` (Optional) whether to print progress to the console.  Default is TRUE.
- `tol` (Optional) parameter change tolerance for stopping training. Default is 1e-20.
- `batch_size` (Optional) change the number of samples used for training 
to reduce computational time. Default is NULL

### Value
Self, i.e., the `torch_lnm` object with updated parameters



## `$lasso_fit()`
Fit a torch_lnm model using the lasso loss function.
This function uses the Adam optimizer to estimate the parameters of a torch_lnm

### Arguments
- `verbose`(Optional) whether to print progress to the console.  Default is TRUE.
- `lrate` (Optional) learning rate of the Adam optimizer. Default is 0.05.
- `maxit` (Optional) maximum number of epochs to train the model. Default is 5000.
- `tol` (Optional) parameter change tolerance for stopping training. Default is 1e-20.
- `v` (Optional) hyperparameter which controls for the penalty term inside the lasso loss function. Default is 1.
- `epsilon` (Optional) Cutoff for lasso to set parameter to 0. Default is 0.0001.
- `batch_size` (Optional) change the number of samples used for training 
to reduce computational time. Default is NULL

### Value
Self, i.e., the `torch_rnm` object with updated parameters



## `$fit()`
Fit a torch_lnm model using a custom loss function supplied to the torch_lnm module.
The custom loss function has to have 2 input parameters, the model covariance matrix and 
the data. (See Example Code for clarification)
This function uses the Adam optimizer to estimate the parameters of a torch_lnm

### Arguments
- `lrate` (Optional) learning rate of the Adam optimizer. Default is 0.05.
- `maxit` (Optional) maximum number of epochs to train the model. Default is 5000.
- `verbose` (Optional) whether to print progress to the console.  Default is TRUE.
- `tol` (Optional) parameter change tolerance for stopping training. Default is 1e-20.
- `batch_size` (Optional) change the number of samples used for training 
to reduce computational time. Default is NULL

### Value
Self, i.e., the `torch_lnm` object with updated parameters



## `$get_df()`
Get the number of degrees of freedom in the model which equals n(n+1)/2 - number of free parameters.
where n is the dimension of the sample covariance matrix.

### Value
Degrees of freedom



## `$get_all_latent_pairings()`
Get all the possible combinations of latent pairs, or the pairs of latent variables as prescribed
in the model. For example, if there are 3 latent variables (a,b,c), we have a total of 3 choose 2, or 3
latent pairs possible and the function will return c('a~b', 'a~c', 'b~c').

### Value
vector of all possible combinations of latent pairs



## `$get_loadings()`
Get all the free factor loadings (entries of lambda matrix) determined after model fit. 
This includes fit, custom_fit and lasso_fit. 

### Value
Dataframe of factor loadings and if the default log-likelihood fn is used in to fit,
standard errors and p-values are also provided



## `$get_residuals()`
Get all the free residuals (diagonal entries of theta matrix) determined after model fit. 
This includes fit, custom_fit and lasso_fit. 

### Value
Dataframe of residuals and if the default log-likelihood fn is used in to fit,
standard errors and p-values are also provided



## `$lasso_update_params_removed(v,epsilon)`
Update the model attribute, self$lasso_num_params_removed. This is will be done automatically if
after `$lasso_fit()`.

### Arguments
- `v` hyperparameter which controls for the penalty term inside the lasso loss function.
- `epsilon`  Cutoff for lasso to set parameter to 0. 


### Value
None



## `$get_partial_correlations(epsilon)`
Get all the free partial correlations (entries of lower triangle of omega_psi matrix) 
determined after model fit. This includes fit, custom_fit and lasso_fit. 

### Arguments
- `epsilon` (Optional) If lasso fit was used, this determines the threshold, under which
a partial correlation is no longer considered a free parameter and is set to 0. Default value is
0.00001



### Value
Dataframe of partial correlations and if the default log-likelihood fn is used in to fit,
standard errors and p-values are also provided



## `$get_criterion_value()`

Get the criterion value of the model according to the type of criterion specified.
The possibilities include AIC (Akaike Information Criterion), 
BIC (Bayesian Information Criterion), EBIC (Extended Bayesian Information Criterion) 
and the chisquare statistic.

### Arguments
- `criterion` Name of the Criterion used, AIC, BIC, EBIC, or chisq.
- `gamma` (optional) Gamma hyperparmaeter for the EBIC if the EBIC is used. (default value is 0.5)

### Value
Criterion Value



## `$get_fit_metrics()`

Get the fit metrics after the model is fit. Metrics returned are the CFI (Comparative Fit Index), 
TLI (Tucker-Lewis Index), the RMSEA (Root Mean Square Error of Approximation) 
and the confidence interval of the RMSEA. These metrics compare the fitted model to
the baseline model which assumes all observed variables are uncorrelated.

### Arguments

### Value
Dataframe containing the fit metrics
}

