% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{torch_jacobian}
\alias{torch_jacobian}
\title{Compute Jacobian of output wrt input tensor}
\usage{
torch_jacobian(output, input)
}
\arguments{
\item{output}{Tensor vector denoting output of target function}

\item{input}{Tensor vector denoting input of target function}
}
\value{
Jacobian matrix
}
\description{
Example: example_function <- function(x) {
y <- x[1]^2 + x[2]^2
z <- x[1] + x[2]
torch_stack(list(y,z))
}
example_tensor <- torch_tensor(c(1,2),requires_grad = TRUE)
example_function(example_tensor)$shape[1]
torch_jacobian(example_function(example_tensor),example_tensor)
}
\details{
Output: 2 x 2 transpose of Jacobian Matrix:
        2 4
        1 1
Explanation: We define an example function (target function)
which outputs an example tensor. Placing both into torch_jacobian
returns the Jacobian of the target function
}
