% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_lnm_helper_functions.R
\name{torch_lnm_stepwise}
\alias{torch_lnm_stepwise}
\alias{tensor_lnm_stepwise}
\title{Torch lnm module created during pruning/stepping up}
\usage{
tensor_lnm_stepwise(params_vec)
}
\arguments{
\item{params_vec}{param values for the initializing of the module}
}
\value{
A `torch_sem` object, which is an `nn_module` (torch object)
}
\description{
Function for creating an lnm model
}
\details{
This function instantiates a torch object for computing the model-implied covariance matrix
based on an lnm model during the pruning/stepping up processes. The methods are identical 
to torch_lnm and information about these methods can be looked up under torch_lnm
}
\seealso{
[df_to_tensor()]
}
