% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_magna_helper_functions.R
\name{transform_omega_params}
\alias{transform_omega_params}
\title{Converts a vector of omega params into a vector containing omega params and zeros along the diagonal
in the original omega matrix}
\usage{
transform_omega_params(omega_params, dim, n)
}
\arguments{
\item{omega_params}{torch tensor or R base vector of omega parameters}

\item{dim}{number of omega parameters}

\item{n}{dimension of network/omega matrix}
}
\value{
torch tensor containing omega parameters and inputed 0s
}
\description{
Example:
transform_omega_params(torch_tensor(c(1,1,1)),3,3)
}
\details{
Output: 
0
1
1
0
1
0

Explanation: zeros are inserted at indices corresponding to the diagonal in the omega matrix, 
where the indices of the omega matrix start from the leftmost column and continue 
downwards and rightwards
}
