% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_lnm_helper_functions.R
\name{vech_dup_idx}
\alias{vech_dup_idx}
\title{Constructs index vector for transforming a vech vector
into a vec vector to create an n*n symmetric matrix
from the vech vector.}
\usage{
vech_dup_idx(n)
}
\arguments{
\item{n}{size of the resulting square matrix}
}
\value{
array containing the indices
}
\description{
Example: idx <-  vech_dup_idx(3) 
        tensor <- torch_tensor(c(1,2,3,4,5,6)) #tensor is a vech vector
        tensor$index_select(0, idx)$view(3,3)
Output: 3 x 3 symmetric matrix:
        1 2 3
        2 4 5
        3 5 6
Explanation: From the vech vector (1,2,3,4,5,6), we from the
vec vector (1,2,3,2,4,5,3,5,6) and from there, create the 
3 x 3 symmetric matrix
}
