% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{vech_to_symmetric_zero_diag}
\alias{vech_to_symmetric_zero_diag}
\title{Transforms vechs vector into an n x n symmetric matrix with 0s along its diagonal}
\usage{
vech_to_symmetric_zero_diag(vech, n)
}
\arguments{
\item{n}{n denoting dimension of output matrix}

\item{x}{vector vechs}
}
\value{
p-value from a 2-tailed Z test
}
\description{
Example: vech <- c(1, 2, 3, 4, 5, 6)
         n <- 4  
         vech_to_symmetric_zero_diag(vech, n)
Output:      
        0    1    2    3
        1    0    4    5
        2    4    0    6
        3    5    6    0
}
\details{
Explanation: We have a vechs of c(1,2,3,4,5,6) which correspond to the lower triangle of our 
output matrix. The diagonal are filled with 0s and the rest of the entries are filled to ensure 
symmetry.
}
