% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torch_lnm_helper_functions.R
\name{lasso_explore}
\alias{lasso_explore}
\title{LASSO for network models}
\usage{
lasso_explore(
  mod,
  criterion = "BIC",
  v_values = NULL,
  epsilon = 1e-05,
  gamma = 0.5
)
}
\arguments{
\item{mod}{original torch_lnm/torch_lnm_stepwise module undergoing stepping up}

\item{criterion}{(Optional) values of the hyperparameter v to be used in the search. Default range 
of values is 30 values of v from 0.01 to 100 spread out on a log scale.}

\item{epsilon}{threshold under which partial correlations will be set to 0}

\item{gamma}{(Optional) value of gamma if "EBIC" is used. Default value is 0.5.}
}
\value{
value of v of the model which gives the best criterion score
}
\description{
Performs a lasso search to find the value of hyperparameter v which correspond to the model 
which gives the lowest criterion scores (AIC/BIC/EBIC). Each model is fit with a different 
value of v and partial correlations which are lower than the threshold value are set to 0/
removed from the pool of free parameters before the criterion score is calculated.
}
