% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{prune}
\alias{prune}
\title{Pruning function for Network Models}
\usage{
prune(mod, batch_size = NULL)
}
\arguments{
\item{mod}{torch module}
}
\value{
new refitted model with only significant partial correlations
}
\description{
Given an  model, it iteratively removes insignificant partial correlations and refits models until
all partial correlations are found to be significant. Thereafter ,it returns this model
with only significant partial correlations.
}
