% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_functions.R
\name{stepup}
\alias{stepup}
\title{Step Up function for Network Models}
\usage{
stepup(mod, criterion = "BIC", gamma = 0.5, batch_size = NULL)
}
\arguments{
\item{mod}{torch module}

\item{criterion}{string indicating criterion to use - "AIC", "BIC or "EBIC"}

\item{gamma}{(Optional) value of gamma if "EBIC" is used. Default value is 0.5.}
}
\value{
either the original torch module or 
another torch module with the best criterion score after stepping down
}
\description{
#' NOT RECOMMENDED FOR RNM  due to long computation times
}
\details{
Performs stepping up to give the best model based on the selected criterion score
At the end of each iteration, one of the originally zero partial correlations will be
turned to non-zero and this corresponds to the model with the best model fit based on criterion score
out of all possibilities (number of zero partial correlations at the start). 
If the models fit at the end of 1 iteration with 1 partial correlation set to non-zero each 
do not lead to a better criterion score, the model found at the end of the previous iteration is returned.

The process terminates if the model is saturated (all partial correlations are non-zero) at the end of the iteration.
}
